+++ ccss/ccss-node-priv.h
@@ -64,10 +64,10 @@ ccss_node_get_style		(ccss_node_t 		*self,
 
 bool
 ccss_node_get_viewport		(ccss_node_t const	*self,
-				 uint32_t		*x,
-				 uint32_t		*y,
-				 uint32_t		*width,
-				 uint32_t		*height);
+				 double			*x,
+				 double			*y,
+				 double			*width,
+				 double			*height);
 
 void
 ccss_node_release		(ccss_node_t		*self);
